% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dosage.R
\name{detect_ploidy}
\alias{detect_ploidy}
\title{Detect Ploidy from Genotype Calls}
\usage{
detect_ploidy(gt_matrix)
}
\arguments{
\item{gt_matrix}{A character matrix of genotype calls (variants x samples).}
}
\value{
A named integer vector of ploidy per sample. Returns the most
common ploidy observed across non-missing genotypes.
}
\description{
Examines a genotype matrix to determine the ploidy level of each sample
based on the number of alleles per genotype call.
}
\examples{
\dontrun{
vcf <- read_vcf("tetraploid.vcf.gz")
gt <- extract_gt(vcf)
detect_ploidy(gt)
# sample1 sample2 sample3
#       4       4       4
}

}
