% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2dosage.R
\name{print.vcf2dosage_result}
\alias{print.vcf2dosage_result}
\title{Print a vcf2dosage Result}
\usage{
\method{print}{vcf2dosage_result}(x, ...)
}
\arguments{
\item{x}{A \code{vcf2dosage_result} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a vcf2dosage Result
}
