% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{print.vcf_data}
\alias{print.vcf_data}
\title{Print a VCF Data Object}
\usage{
\method{print}{vcf_data}(x, ...)
}
\arguments{
\item{x}{A \code{vcf_data} object.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a VCF Data Object
}
