% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_vcf}
\alias{validate_vcf}
\title{Validate a VCF Data Object}
\usage{
validate_vcf(vcf)
}
\arguments{
\item{vcf}{Object to validate.}
}
\value{
Invisible NULL. Throws error if invalid.
}
\description{
Validate a VCF Data Object
}
\keyword{internal}
