% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vcf2dosage-package.R
\docType{package}
\name{vcf2dosage-package}
\alias{vcf2dosage-package}
\title{vcf2dosage: Polyploid-Aware VCF to Dosage Matrix Conversion for Genomic Selection}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Converts Variant Call Format (VCF) files to numeric allele dosage matrices ready for genomic selection and genome-wide association studies. Handles diploid and polyploid species (tetraploid, hexaploid, and arbitrary ploidy), filters variants by minor allele frequency, missingness, and read depth, and exports dosage matrices in formats compatible with 'rrBLUP', 'GWASpoly', 'BGLR', 'sommer', and 'AGHmatrix'. Supports both unphased (0/0/1/1) and phased (0|0|1|1) genotype calls.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/josh45-source/vcf2dosage}
  \item Report bugs at \url{https://github.com/josh45-source/vcf2dosage/issues}
}

}
\author{
\strong{Maintainer}: Joash Joshua Ayo \email{joash.ayo@example.com}

}
\keyword{internal}
